+++ truck-meshalgo/src/tessellation/triangulation.rs
@@ -37,7 +37,9 @@ where
     S: RobustMeshableSurface,
 {
     surface
-        .search_nearest_parameter(point, hint, 100)
+        .search_parameter(point, hint, 100)
+        .or_else(|| surface.search_parameter(point, None, 100))
+        .or_else(|| surface.search_nearest_parameter(point, hint, 100))
         .or_else(|| surface.search_nearest_parameter(point, None, 100))
 }